backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
bcrypt==4.1.2
# Testing
pytest==8.3.2
pytest-flask==1.3.0
